rename memory view -> memory gauge view
