+++ lib/dashboard.js
@@ -5,7 +5,8 @@ var blessed = require("blessed");
 
 var StreamView = require("./views/stream-view");
 var EventLoopView = require("./views/eventloop-view");
-var MemoryView = require("./views/memory-view");
+var MemoryGaugeView = require("./views/memory-gauge-view");
+var MemoryGraphView = require("./views/memory-graph-view");
 var CpuView = require("./views/cpu-view");
 var HelpView = require("./views/help");
 var generateLayouts = require("./generate-layouts");
@@ -88,7 +89,7 @@ Dashboard.prototype.onEvent = function (event) {
 var VIEW_MAP = {
   log: StreamView,
   cpu: CpuView,
-  memory: MemoryView,
+  memory: MemoryGaugeView,
   eventLoop: EventLoopView
 };
