+++ lib/views/memory-gauge-view.js
